requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
numpy==1.26.2
pypdf==3.17.1
